RHEL8: package task instead of yum
Switch from 'yum' to 'package' task to be compatible with dnf.

In cases when it's used for listing,
it could happen such role/part this usage is in,
would be run as non-root user, but that may not work
if python*-dnf pkg is not present on the system.

When this happens, package task does attempt to auto-install it
and fail when not being root.
So ensure that in package: list ... cases, we explicitly
become root (for package: install cases that is implicitly required).

Change-Id: I963ef2c65e8e45e7f1cf6314a3f532caee45cebe